src/runner.py
# src/runner.py
import argparse
import yaml

def main(config):
    print("✅ Runner started with config:")
    print(config)

if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("--config", type=str, required=True)
    args = parser.parse_args()

    with open(args.config, "r") as f:
        config = yaml.safe_load(f)

    main(config)
