requirements.txt
torch
torchvision
numpy
pandas
matplotlib
tqdm
pyyaml
